+++ test/package-api.ts
@@ -1,4 +1,4 @@
-import { type LintOptions, lint, type Report } from 'amigolint';
+import { type LintOptions, lint, type Report } from 'lintamigo';
 
 const lintFunction: (options: LintOptions) => Promise<Report> = lint;
 void lintFunction;